getirdeki urunleri alt kategorileri ile birlikte çekebilyior

+++ python/model/Price.py
@@ -4,7 +4,7 @@ class Price:
                 id : int = None,
                 price_value : float = None,
                 product_id : int = None,
-                time_unix : int = None) -> None:
+                time_unix : float = None) -> None:
         self.id = id
         self.price_value = price_value
         self.product_id = product_id
